Split order creation and batch finalization out of the transaction preparation service
TransactionOrderFactory now owns order construction with quantity resolution
and price-staleness checks; TransactionBatchFinalizer owns confirmed-batch
finalization with the after-commit Drive upload, custodian email, and event
publishing, keeping the synchronization registration inside the same
transactional call stack as before. The preparation service keeps only
processCommand orchestration and audit payload building, and the admin
service and command job now inject the finalizer directly.

God classes 4 -> 3, PMD violations 133 -> 131, top-15 class lines 6478 -> 6305.
