+++ metrics/scorecard.json
@@ -1,5 +1,5 @@
 {
-  "branchCoverage": 87.67,
+  "branchCoverage": 87.68,
   "classCouplingViolations": 34,
   "clockViolations": 0,
   "cognitiveComplexityViolations": 20,
@@ -8,14 +8,14 @@
   "disabledTests": 5,
   "disconnectedClasses": 452,
   "fqnViolations": 0,
-  "godClasses": 4,
-  "interactionVerifyingTestClasses": 362,
+  "godClasses": 3,
+  "interactionVerifyingTestClasses": 363,
   "interactionVerifyingTestShare": 36.1,
   "lineCoverage": 84.04,
   "longClasses": 1,
   "meanInstability": 0.473,
   "mocklessTestClasses": 462,
-  "mocklessTestShare": 46.1,
+  "mocklessTestShare": 46.0,
   "moduleCycles": 0,
   "moduleSliceTests": 0,
   "modulithViolations": 0,
@@ -23,26 +23,26 @@
   "mutationScoreSliced": 89.81,
   "mutationsSliced": 13028,
   "mutationsTotal": 13028,
-  "overMockedTestClasses": 111,
-  "overMockedTestShare": 11.1,
+  "overMockedTestClasses": 113,
+  "overMockedTestShare": 11.2,
   "pitestSlicedModules": 52,
-  "pmdViolations": 133,
+  "pmdViolations": 131,
   "pmdViolationsByRule": {
     "CognitiveComplexity": 20,
     "CouplingBetweenObjects": 34,
-    "CyclomaticComplexity": 71,
-    "GodClass": 4,
+    "CyclomaticComplexity": 70,
+    "GodClass": 3,
     "TooManyMethods": 4
   },
-  "slowest10TestClassesSeconds": 114.9,
+  "slowest10TestClassesSeconds": 117.2,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 164,
-  "testClassesOver10Seconds": 4,
-  "testClassesTotal": 1003,
+  "suiteWallTimeSeconds": 201,
+  "testClassesOver10Seconds": 5,
+  "testClassesTotal": 1005,
   "testIntegrationShare": 6.7,
-  "testSliceShare": 13.8,
+  "testSliceShare": 13.7,
   "testUnitShare": 79.6,
-  "top15ClassLines": 6478,
+  "top15ClassLines": 6305,
   "totalPackages": 227,
   "unmarkedPackages": 0
 }
