+++ src/main/java/ee/tuleva/onboarding/investment/transaction/TransactionPreparationService.java
@@ -1,50 +1,23 @@
 package ee.tuleva.onboarding.investment.transaction;
 
-import static ee.tuleva.onboarding.investment.JobRunSchedule.TIMEZONE;
-import static java.math.BigDecimal.ZERO;
-import static java.util.stream.Collectors.toMap;
-
-import ee.tuleva.onboarding.comparisons.fundvalue.PositionPriceResolver;
-import ee.tuleva.onboarding.comparisons.fundvalue.ResolvedPrice;
 import ee.tuleva.onboarding.investment.epis.SettlementTimingWarningService;
-import ee.tuleva.onboarding.investment.portfolio.ModelPortfolioAllocation;
-import ee.tuleva.onboarding.investment.portfolio.ModelPortfolioAllocationRepository;
 import ee.tuleva.onboarding.investment.transaction.calculation.TradeCalculationEngine;
-import ee.tuleva.onboarding.investment.transaction.export.CustodianOrderEmailSender;
-import ee.tuleva.onboarding.investment.transaction.export.GoogleDriveProperties;
-import ee.tuleva.onboarding.investment.transaction.export.TransactionExportService;
-import ee.tuleva.onboarding.investment.transaction.export.TransactionExportUploader;
 import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.time.Clock;
 import java.time.Instant;
-import java.time.LocalDate;
-import java.time.ZoneId;
-import java.time.temporal.ChronoUnit;
-import java.util.ArrayList;
-import java.util.Base64;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
-import java.util.function.Function;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.jspecify.annotations.Nullable;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionSynchronization;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class TransactionPreparationService {
-
-  private static final int STALE_PRICE_THRESHOLD_DAYS = 3;
-  private static final ZoneId TALLINN = ZoneId.of(TIMEZONE);
 
   private final TransactionInputService inputService;
   private final TradeCalculationEngine calculationEngine;
@@ -52,14 +25,7 @@ public class TransactionPreparationService {
   private final TransactionOrderRepository orderRepository;
   private final TransactionAuditEventRepository auditEventRepository;
   private final TransactionCommandRepository commandRepository;
-  private final SettlementDateCalculator settlementDateCalculator;
-  private final TransactionExportService exportService;
-  private final ModelPortfolioAllocationRepository modelPortfolioAllocationRepository;
-  private final PositionPriceResolver positionPriceResolver;
-  private final ApplicationEventPublisher eventPublisher;
-  private final GoogleDriveProperties driveProperties;
-  @Nullable private final TransactionExportUploader exportUploader;
-  private final CustodianOrderEmailSender custodianOrderEmailSender;
+  private final TransactionOrderFactory orderFactory;
   private final SettlementTimingWarningService settlementTimingWarningService;
   private final Clock clock;
 
@@ -89,7 +55,7 @@ public class TransactionPreparationService {
       batchRepository.save(batch);
 
       CalculatedOrders calculated =
-          createOrders(batch, result, command.getAsOfDate(), Instant.now(clock));
+          orderFactory.createOrders(batch, result, command.getAsOfDate(), Instant.now(clock));
       List<TransactionOrder> orders = calculated.orders();
       orderRepository.saveAll(orders);
 
@@ -147,10 +113,6 @@ public class TransactionPreparationService {
     return command.getActor() == null ? "system" : command.getActor();
   }
 
-  private static String finalizingActor(TransactionBatch batch) {
-    return batch.getConfirmedBy() == null ? "system" : batch.getConfirmedBy();
-  }
-
   private Map<String, Object> failedPayload(
       TransactionCommand command, @Nullable FundTransactionInput input, RuntimeException e) {
     Map<String, Object> payload = new LinkedHashMap<>();
@@ -164,254 +126,6 @@ public class TransactionPreparationService {
     payload.put("exceptionClass", e.getClass().getName());
     putIfPresent(payload, "errorMessage", command.getErrorMessage());
     return payload;
-  }
-
-  @Transactional
-  public void finalizeConfirmedBatch(TransactionBatch batch) {
-    if (batch.getStatus() == BatchStatus.SENT) {
-      throw new IllegalStateException(
-          "Batch already finalized: id=" + batch.getId() + ", status=" + batch.getStatus());
-    }
-    log.info("Finalizing batch: id={}", batch.getId());
-
-    Instant now = Instant.now(clock);
-    LocalDate tradeDate = now.atZone(TALLINN).toLocalDate();
-
-    List<TransactionOrder> orders = orderRepository.findByBatchId(batch.getId());
-    requireQuantitiesForNonAmountOrders(batch, orders);
-    orders.forEach(
-        order -> {
-          order.setOrderTimestamp(now);
-          order.setExpectedSettlementDate(
-              settlementDateCalculator.calculateSettlementDate(
-                  now, order.getInstrumentType(), order.getInstrumentIsin()));
-          order.setOrderStatus(OrderStatus.SENT);
-        });
-    orderRepository.saveAll(orders);
-
-    List<ModelPortfolioAllocation> currentAllocations =
-        modelPortfolioAllocationRepository.findLatestByFundAsOf(batch.getFund(), tradeDate);
-    List<ModelPortfolioAllocation> previousAllocations =
-        modelPortfolioAllocationRepository.findPreviousByFundAsOf(batch.getFund(), tradeDate);
-    var mergedAllocations = new ArrayList<>(previousAllocations);
-    mergedAllocations.addAll(currentAllocations);
-    Map<String, String> labelsByIsin =
-        buildLookupMap(mergedAllocations, ModelPortfolioAllocation::getLabel);
-    Map<String, String> ricByIsin =
-        buildLookupMap(mergedAllocations, ModelPortfolioAllocation::getTicker);
-    Map<String, String> bbgByIsin =
-        buildLookupMap(mergedAllocations, ModelPortfolioAllocation::getBbgTicker);
-
-    byte[] xlsxExport = exportService.generateOrdersExport(orders);
-    byte[] sebFundXlsx = exportService.generateSebFundExport(orders, labelsByIsin);
-    byte[] sebEtfXlsx = exportService.generateSebEtfExport(orders, ricByIsin);
-    byte[] ftEtfXlsx =
-        exportService.generateFtEtfExport(orders, labelsByIsin, bbgByIsin, tradeDate);
-    byte[] uuidWorkbookXlsx = exportService.generateUuidWorkbook(orders);
-
-    Map<String, Object> updatedMetadata = new HashMap<>(batch.getMetadata());
-    updatedMetadata.put("xlsxExport", encodeExport(xlsxExport));
-    updatedMetadata.put("sebFundXlsx", encodeExport(sebFundXlsx));
-    updatedMetadata.put("sebEtfXlsx", encodeExport(sebEtfXlsx));
-    updatedMetadata.put("ftEtfXlsx", encodeExport(ftEtfXlsx));
-    updatedMetadata.put("uuidWorkbookXlsx", encodeExport(uuidWorkbookXlsx));
-    batch.setMetadata(updatedMetadata);
-
-    batch.setStatus(BatchStatus.SENT);
-    batchRepository.save(batch);
-
-    auditEventRepository.save(
-        TransactionAuditEvent.builder()
-            .batch(batch)
-            .eventType("BATCH_FINALIZED")
-            .actor(finalizingActor(batch))
-            .createdAt(Instant.now(clock))
-            .payload(Map.of("tradeDate", tradeDate.toString(), "orderCount", orders.size()))
-            .build());
-
-    runAfterCommit(
-        () ->
-            publishExportsToDrive(
-                batch,
-                now,
-                tradeDate,
-                orders.size(),
-                sebFundXlsx,
-                sebEtfXlsx,
-                ftEtfXlsx,
-                uuidWorkbookXlsx));
-
-    log.info("Batch finalized: id={}, orderCount={}", batch.getId(), orders.size());
-  }
-
-  private void publishExportsToDrive(
-      TransactionBatch batch,
-      Instant timestamp,
-      LocalDate tradeDate,
-      int orderCount,
-      byte[] sebFundXlsx,
-      byte[] sebEtfXlsx,
-      byte[] ftEtfXlsx,
-      byte[] uuidWorkbookXlsx) {
-    var exports =
-        Map.of(
-            "sebFundXlsx", sebFundXlsx,
-            "sebEtfXlsx", sebEtfXlsx,
-            "ftEtfXlsx", ftEtfXlsx,
-            "uuidWorkbookXlsx", uuidWorkbookXlsx);
-    Map<String, String> driveFileUrls = uploadExportsToDrive(batch, timestamp, exports);
-    if (!driveFileUrls.isEmpty()) {
-      persistDriveFileUrls(batch, driveFileUrls);
-    }
-    custodianOrderEmailSender.send(batch.getFund(), timestamp, exports);
-    eventPublisher.publishEvent(
-        new BatchFinalizedEvent(batch.getId(), orderCount, tradeDate.toString(), driveFileUrls));
-  }
-
-  void persistDriveFileUrls(TransactionBatch batch, Map<String, String> driveFileUrls) {
-    Map<String, Object> updatedMetadata = new HashMap<>(batch.getMetadata());
-    updatedMetadata.put("driveFileUrls", driveFileUrls);
-    batch.setMetadata(updatedMetadata);
-    batchRepository.save(batch);
-  }
-
-  private void runAfterCommit(Runnable action) {
-    if (TransactionSynchronizationManager.isSynchronizationActive()) {
-      TransactionSynchronizationManager.registerSynchronization(
-          new TransactionSynchronization() {
-            @Override
-            public void afterCommit() {
-              action.run();
-            }
-          });
-    } else {
-      action.run();
-    }
-  }
-
-  private CalculatedOrders createOrders(
-      TransactionBatch batch, FundCalculationResult result, LocalDate asOfDate, Instant createdAt) {
-    var input = result.input();
-    List<TransactionOrder> orders = new ArrayList<>();
-    Map<String, @Nullable ResolvedPrice> priceResolutions = new LinkedHashMap<>();
-
-    for (TradeCalculation trade : result.trades()) {
-      if (trade.tradeAmount().compareTo(ZERO) == 0) {
-        continue;
-      }
-      String isin =
-          Objects.requireNonNull(
-              trade.isin(), "Trade with non-zero amount is missing isin: fund=" + result.fund());
-      var instrumentType = input.instrumentTypes().getOrDefault(isin, InstrumentType.ETF);
-      var transactionType =
-          trade.tradeAmount().compareTo(ZERO) > 0 ? TransactionType.BUY : TransactionType.SELL;
-      var orderAmount = trade.tradeAmount().abs();
-      var orderQuantity =
-          resolveOrderQuantity(instrumentType, transactionType, isin, orderAmount, asOfDate);
-      if (!isAmountBasedOrder(instrumentType, transactionType)) {
-        priceResolutions.put(isin, orderQuantity.resolvedPrice());
-      }
-      orders.add(
-          TransactionOrder.builder()
-              .batch(batch)
-              .fund(result.fund())
-              .instrumentIsin(isin)
-              .transactionType(transactionType)
-              .instrumentType(instrumentType)
-              .orderAmount(orderAmount)
-              .orderQuantity(orderQuantity.quantity())
-              .comment(orderQuantity.stalePriceComment())
-              .orderVenue(input.orderVenues().getOrDefault(isin, OrderVenue.SEB))
-              .createdAt(createdAt)
-              .build());
-    }
-    return new CalculatedOrders(orders, priceResolutions);
-  }
-
-  private OrderQuantity resolveOrderQuantity(
-      InstrumentType instrumentType,
-      TransactionType transactionType,
-      String isin,
-      BigDecimal orderAmount,
-      LocalDate asOfDate) {
-    if (isAmountBasedOrder(instrumentType, transactionType)) {
-      return new OrderQuantity(null, null, null);
-    }
-    ResolvedPrice resolvedPrice = positionPriceResolver.resolve(isin, asOfDate).orElse(null);
-    if (resolvedPrice == null) {
-      log.warn(
-          "No price found for order quantity: isin={}, instrumentType={}, asOfDate={}",
-          isin,
-          instrumentType,
-          asOfDate);
-      return new OrderQuantity(null, null, null);
-    }
-    BigDecimal price = resolvedPrice.usedPrice();
-    if (price == null) {
-      log.warn(
-          "No price found for order quantity: isin={}, instrumentType={}, asOfDate={}",
-          isin,
-          instrumentType,
-          asOfDate);
-      return new OrderQuantity(null, null, resolvedPrice);
-    }
-    if (price.signum() <= 0) {
-      log.warn(
-          "Non-positive price for order quantity, leaving quantity unset:"
-              + " isin={}, instrumentType={}, price={}, asOfDate={}",
-          isin,
-          instrumentType,
-          price.toPlainString(),
-          asOfDate);
-      return new OrderQuantity(null, null, resolvedPrice);
-    }
-    BigDecimal quantity = orderAmount.divide(price, 6, RoundingMode.HALF_UP);
-    String stalePriceComment = describeIfStale(isin, resolvedPrice, asOfDate);
-    return new OrderQuantity(quantity, stalePriceComment, resolvedPrice);
-  }
-
-  @Nullable
-  private String describeIfStale(String isin, ResolvedPrice resolvedPrice, LocalDate asOfDate) {
-    LocalDate priceDate = resolvedPrice.priceDate();
-    if (priceDate == null) {
-      return null;
-    }
-    long ageDays = ChronoUnit.DAYS.between(priceDate, asOfDate);
-    if (ageDays <= STALE_PRICE_THRESHOLD_DAYS) {
-      return null;
-    }
-    log.warn(
-        "Order sized on stale price: isin={}, priceDate={}, ageDays={}, source={}, asOfDate={}",
-        isin,
-        priceDate,
-        ageDays,
-        resolvedPrice.priceSource(),
-        asOfDate);
-    return "Sized on stale price: priceDate=%s, ageDays=%d, source=%s"
-        .formatted(priceDate, ageDays, resolvedPrice.priceSource());
-  }
-
-  private boolean isAmountBasedOrder(
-      InstrumentType instrumentType, TransactionType transactionType) {
-    return instrumentType == InstrumentType.FUND && transactionType == TransactionType.BUY;
-  }
-
-  private void requireQuantitiesForNonAmountOrders(
-      TransactionBatch batch, List<TransactionOrder> orders) {
-    List<String> missingQuantity =
-        orders.stream()
-            .filter(
-                order -> !isAmountBasedOrder(order.getInstrumentType(), order.getTransactionType()))
-            .filter(order -> order.getOrderQuantity() == null)
-            .map(TransactionOrder::getInstrumentIsin)
-            .toList();
-    if (!missingQuantity.isEmpty()) {
-      throw new IllegalStateException(
-          "Cannot finalize batch: orders require a quantity but have none: batchId="
-              + batch.getId()
-              + ", isins="
-              + missingQuantity);
-    }
   }
 
   private Map<String, Object> completedPayload(
@@ -464,32 +178,5 @@ public class TransactionPreparationService {
   @Nullable
   private static String plain(@Nullable BigDecimal value) {
     return value == null ? null : value.toPlainString();
-  }
-
-  private Map<String, String> buildLookupMap(
-      List<ModelPortfolioAllocation> allocations,
-      Function<ModelPortfolioAllocation, @Nullable String> valueExtractor) {
-    return allocations.stream()
-        .filter(
-            allocation -> allocation.getIsin() != null && valueExtractor.apply(allocation) != null)
-        .collect(toMap(ModelPortfolioAllocation::getIsin, valueExtractor, (a, b) -> b));
-  }
-
-  private Map<String, String> uploadExportsToDrive(
-      TransactionBatch batch, Instant timestamp, Map<String, byte[]> exports) {
-    if (!driveProperties.enabled() || exportUploader == null) {
-      return Map.of();
-    }
-    try {
-      return exportUploader.uploadExports(
-          driveProperties.rootFolderId(), batch.getFund(), timestamp, exports);
-    } catch (Exception e) {
-      log.error("Google Drive upload failed: batchId={}", batch.getId(), e);
-      return Map.of();
-    }
-  }
-
-  private String encodeExport(byte[] export) {
-    return Base64.getEncoder().encodeToString(export);
   }
 }
